backend/requirements.txt
fastapi>=0.135.0
uvicorn[standard]>=0.42.0
aiosqlite>=0.22.0
aiofiles>=25.1.0
python-multipart>=0.0.22
rembg[cpu]>=2.0.74
Pillow>=12.0.0
onnxruntime>=1.24.0
pydantic>=2.12.0
pydantic-settings>=2.13.0
python-dotenv>=1.2.0
